implement an centralized PydanticParsingSettings class
