+++ src/geh_common/telemetry/pydantic_settings_parsing.py
@@ -1,0 +1,38 @@
+from typing import Tuple, Type
+
+from pydantic_settings import (
+    BaseSettings,
+    CliSettingsSource,
+    PydanticBaseSettingsSource
+)
+
+
+class PydanticParsingSettings(
+    BaseSettings,
+    cli_parse_args=True,
+    cli_kebab_case=True,
+    cli_ignore_unknown_args=True,
+    cli_implicit_flags=True,
+):
+    """
+    Base class for application settings.
+
+    Supports:
+    - CLI parsing with arguments using kebab-case.
+    - Environment variables using SNAKE_UPPER_CASE.
+    - Ignoring unknown CLI arguments. This behavior can be overridden by setting `cli_ignore_unknown_args=False`
+      in the class definition of the derived settings class. Example:
+      `class LoggingSettings(ApplicationSettings, cli_ignore_unknown_args=False):`
+    """
+
+    @classmethod
+    def settings_customise_sources(
+        cls,
+        settings_cls: Type[BaseSettings],
+        init_settings: PydanticBaseSettingsSource,
+        env_settings: PydanticBaseSettingsSource,
+        dotenv_settings: PydanticBaseSettingsSource,
+        file_secret_settings: PydanticBaseSettingsSource,
+    ) -> Tuple[PydanticBaseSettingsSource, ...]:
+        """Determines the priority of loading field values in the returned order of settings"""
+        return CliSettingsSource(settings_cls), env_settings, init_settings
